refactor vendor ui package name
