+++ app/src/main/java/com/example/myapplication/vendor_ui/account/AccountFragment.java
@@ -1,18 +1,14 @@
-package com.example.myapplication.ui.account;
+package com.example.myapplication.vendor_ui.account;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 
 import com.example.myapplication.HomeActivity;
