+++ app/src/main/res/layout/fragment_batch_invoice.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.BatchInvoiceFragment">
+    tools:context=".vendor_ui.fragments.BatchInvoiceFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
