+++ app/src/main/res/layout/fragment_saved.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.saved.SavedFragment">
+    tools:context=".vendor_ui.saved.SavedFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/savedToolbar"
