+++ app/src/main/res/layout/fragment_transactions.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.transactions.TransactionsFragment">
+    tools:context=".vendor_ui.transactions.TransactionsFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/transactionsToolbar"
